test(ms): storage - simplify test in one file
